Move primitives-related tests
